Move main hermes script from __init__.py
